Generated overview map
Generated a figure that displays Florida and the Osceola national forest
shapefiles. In doing so, wrote a function that reads in a shapefile just
to extract the polygon bounds for later referencing in a basemap
instance.

A side note, the vector geometry for the Osceola that I received was
haneously overcomplicated. I manually removed unnecessary vertices and
vertices that actually broke continuity in the geometry.