Included mixed pine IC rasters and text files
Completed the generation of the mixed pine initial communities raster,
generated via a raster of stand ages. Also created a function to write a
list of dicts to a text file, where each dict entry is in the form of a
LANDIS-II mapcode, and ahve the text file be formatted correctly for an
initial communities input.