Finished swamp FIA plot stuff
and have working IC assignment code for swamp ecoregions. This includes
some code to identify that region the IC generator is currently in, if
that cell is an 'edge' and is adjacent to another region, and if the
region has been previously assigned a map code (if so, it grabs the age
data and uses that to drive the cohort info).